Challange Day-34 : One of the best ways to learn is to write your own code as often as possible, so here are three ways you should try extending this app to make sure you fully understand what’s going on.
Go back to the Guess the Flag project and add some animation:

When you tap a flag, make it spin around 360 degrees on the Y axis.
Make the other two buttons fade out to 25% opacity.
Add a third effect of your choosing to the two flags the user didn’t choose – maybe make them scale down? Or flip in a different direction? Experiment!